Signal finish

+++ srcs/signal/sh_signal.c
@@ -6,7 +6,7 @@
 /*   By: eboris <eboris@student.21-school.ru>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/06 17:27:09 by eboris            #+#    #+#             */
-/*   Updated: 2020/11/19 16:47:06 by eboris           ###   ########.fr       */
+/*   Updated: 2020/11/20 15:55:00 by eboris           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,14 +20,17 @@ void	sh_signal_set(void)
 
 void	sh_signal_status(int status, pid_t pid)
 {
-	kill(pid, 1);
-	if (status == SIGSEGV)
-		ft_fprintf(STDERR_FILENO, "Segmentation fault: ");
-	else if (status == SIGBUS)
-		ft_fprintf(STDERR_FILENO, "Bus error: ");
-	else if (status == SIGABRT)
-		ft_fprintf(STDERR_FILENO, "Abort trap: ");
-	else if (status == SIGFPE)
-		ft_fprintf(STDERR_FILENO, "Floating point exception: ");
-	ft_fprintf(STDERR_FILENO, "Status code: %i\n", status);
+	if (!WIFEXITED(status))
+	{
+		kill(pid, 1);
+		if (status == SIGSEGV)
+			ft_fprintf(STDERR_FILENO, "Segmentation fault: ");
+		else if (status == SIGBUS)
+			ft_fprintf(STDERR_FILENO, "Bus error: ");
+		else if (status == SIGABRT)
+			ft_fprintf(STDERR_FILENO, "Abort trap: ");
+		else if (status == SIGFPE)
+			ft_fprintf(STDERR_FILENO, "Floating point exception: ");
+		ft_fprintf(STDERR_FILENO, "Status code: %i\n", status);
+	}
 }
